修复断线重连后出生点的问题

+++ src/main/java/cn/nukkit/Player.java
@@ -2252,7 +2252,7 @@ public class Player extends EntityHuman implements CommandSender, InventoryHolde
 
         nbt.putString("NameTag", this.username);
 
-        if (0 >= nbt.getShort("Health")) {
+        if (0 >= nbt.getFloat("Health")) {
             alive = false;
         }
